+++ examples/storybook/src/shared/resources/experimentsExample.ts
@@ -16,7 +16,6 @@ export const EXPERIMENTS = [
   { name: 'anchorScrollWithoutScrollIntoView', scope: 'ricos' },
   { name: 'anchorScrollWithSmoothScrollIntoViewIfNeeded', scope: 'ricos' },
   { name: 'useSrcSet', scope: 'ricos' },
-  { name: 'pastedFilesSupport', scope: 'ricos' },
   { name: 'normalizePoll', scope: 'ricos' },
   { name: 'optimizeFullScreenModal', scope: 'ricos' },
   { name: 'useFilePluginAutoDownloadLinkRef', scope: 'ricos' },